# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinBot API",
    "description": "Financial onboarding and transaction analysis. Generated from the route Zod schemas; regenerate with `npm run openapi:generate`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/onboarding/manual": {
      "put": {
        "operationId": "saveManualOnboarding",
        "summary": "Save manual answers; completes the manual gate only",
        "responses": {
          "200": {
            "description": "Saved"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "dateOfBirth": {
                    "type": "string"
                  },
                  "maritalStatus": {
                    "type": "string",
                    "enum": [
                      "single",
                      "married",
                      "domestic_partnership",
                      "divorced",
                      "widowed",
                      "prefer_not_to_say"
                    ]
                  },
                  "dependentsCount": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "employmentStatus": {
                    "type": "string",
                    "enum": [
                      "full_time",
                      "part_time",
                      "self_employed",
                      "unemployed",
                      "retired",
                      "student"
                    ]
                  },
                  "monthlyTakeHomeIncome": {
                    "type": "number",
                    "minimum": 0
                  },
                  "monthlyHousingCosts": {
                    "type": "number",
                    "minimum": 0
                  },
                  "monthlyFoodSpend": {
                    "type": "number",
                    "minimum": 0
                  },
                  "monthlyTransportationCosts": {
                    "type": "number",
                    "minimum": 0
                  },
                  "subscriptions": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "netflix",
                        "disney_hulu",
                        "amazon_prime",
                        "paramount",
                        "apple",
                        "xbox",
                        "playstation",
                        "nintendo",
                        "none"
                      ]
                    }
                  },
                  "monthlyEntertainmentSubscriptionsCosts": {
                    "type": "number",
                    "minimum": 0
                  },
                  "savingsAndEmergencyFunds": {
                    "type": "number",
                    "minimum": 0
                  },
                  "totalDebt": {
                    "type": "number",
                    "minimum": 0
                  },
                  "factorInDebtInterest": {
                    "type": "boolean"
                  },
                  "financialGoals": {
                    "minItems": 3,
                    "maxItems": 3,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "emergency_fund",
                        "pay_off_debt",
                        "save_for_retirement",
                        "save_for_home",
                        "invest_more",
                        "reduce_spending"
                      ]
                    }
                  },
                  "additionalMoneyPools": {
                    "minItems": 3,
                    "maxItems": 3,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "vacation",
                        "miscellaneous",
                        "emergency",
                        "savings",
                        "investing"
                      ]
                    }
                  },
                  "riskComfort": {
                    "type": "string",
                    "enum": [
                      "conservative",
                      "moderate",
                      "aggressive"
                    ]
                  },
                  "additionalContext": {
                    "type": "string"
                  }
                },
                "required": [
                  "fullName",
                  "dateOfBirth",
                  "maritalStatus",
                  "dependentsCount",
                  "employmentStatus",
                  "monthlyTakeHomeIncome",
                  "monthlyHousingCosts",
                  "monthlyFoodSpend",
                  "monthlyTransportationCosts",
                  "subscriptions",
                  "savingsAndEmergencyFunds",
                  "totalDebt",
                  "factorInDebtInterest",
                  "financialGoals",
                  "additionalMoneyPools",
                  "riskComfort",
                  "additionalContext"
                ]
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "getManualOnboarding",
        "summary": "Resume saved manual answers",
        "responses": {
          "200": {
            "description": "Saved answers or null",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "saved": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "payload": {
                              "type": "object",
                              "propertyNames": {
                                "type": "string"
                              },
                              "additionalProperties": {}
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "payload",
                            "updatedAt"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "saved"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/onboarding/status": {
      "get": {
        "operationId": "getOnboardingStatus",
        "summary": "Gates, phase, institution progress, and available actions",
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "phase": {
                      "type": "string",
                      "enum": [
                        "financial_linking",
                        "manual_profile_in_progress",
                        "waiting_for_history",
                        "classifying",
                        "review_ready",
                        "recomputing",
                        "failed_retryable",
                        "complete"
                      ]
                    },
                    "gates": {
                      "type": "object",
                      "properties": {
                        "hasLinkedInstitution": {
                          "type": "boolean"
                        },
                        "linkingDeclaredComplete": {
                          "type": "boolean"
                        },
                        "manualProfileComplete": {
                          "type": "boolean"
                        },
                        "analysisReviewable": {
                          "type": "boolean"
                        },
                        "financialReviewConfirmed": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "hasLinkedInstitution",
                        "linkingDeclaredComplete",
                        "manualProfileComplete",
                        "analysisReviewable",
                        "financialReviewConfirmed"
                      ],
                      "additionalProperties": false
                    },
                    "analysis": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "runId": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "waiting_for_history",
                                "processing",
                                "review_ready",
                                "recomputing",
                                "confirmed",
                                "failed",
                                "superseded"
                              ]
                            },
                            "requestedLookbackDays": {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            "institutions": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "integer",
                                  "minimum": -9007199254740991,
                                  "maximum": 9007199254740991
                                },
                                "ready": {
                                  "type": "integer",
                                  "minimum": -9007199254740991,
                                  "maximum": 9007199254740991
                                },
                                "limited": {
                                  "type": "integer",
                                  "minimum": -9007199254740991,
                                  "maximum": 9007199254740991
                                },
                                "failed": {
                                  "type": "integer",
                                  "minimum": -9007199254740991,
                                  "maximum": 9007199254740991
                                },
                                "pending": {
                                  "type": "integer",
                                  "minimum": -9007199254740991,
                                  "maximum": 9007199254740991
                                }
                              },
                              "required": [
                                "total",
                                "ready",
                                "limited",
                                "failed",
                                "pending"
                              ],
                              "additionalProperties": false
                            },
                            "startedAt": {
                              "type": "string"
                            },
                            "reviewReadyAt": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            },
                            "retryAllowed": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "runId",
                            "status",
                            "requestedLookbackDays",
                            "institutions",
                            "startedAt",
                            "reviewReadyAt",
                            "retryAllowed"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "availableActions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "onboardingComplete": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "phase",
                    "gates",
                    "analysis",
                    "availableActions",
                    "onboardingComplete"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "phase": "waiting_for_history",
                  "gates": {
                    "hasLinkedInstitution": true,
                    "linkingDeclaredComplete": true,
                    "manualProfileComplete": true,
                    "analysisReviewable": false,
                    "financialReviewConfirmed": false
                  },
                  "analysis": {
                    "runId": "5b910c92-1890-4dce-8912-0e496d4091a4",
                    "status": "waiting_for_history",
                    "requestedLookbackDays": 180,
                    "institutions": {
                      "total": 2,
                      "ready": 1,
                      "limited": 0,
                      "failed": 0,
                      "pending": 1
                    },
                    "startedAt": "2026-08-24T20:00:00Z",
                    "reviewReadyAt": null,
                    "retryAllowed": false
                  },
                  "availableActions": [
                    "view_waiting",
                    "manage_connections",
                    "manage_notifications",
                    "logout"
                  ],
                  "onboardingComplete": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/onboarding/financial-review/recompute": {
      "post": {
        "operationId": "recomputeReview",
        "summary": "Rebuild facts and review after corrections",
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "already_recomputing"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "ANALYSIS_NOT_REVIEWABLE",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/onboarding/financial-review/confirm": {
      "post": {
        "operationId": "confirmReview",
        "summary": "Confirm the latest review; the only path to completion",
        "responses": {
          "200": {
            "description": "Confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "onboardingComplete": {
                      "type": "boolean"
                    },
                    "alreadyConfirmed": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "onboardingComplete",
                    "alreadyConfirmed"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "REVIEW_VERSION_STALE or REVIEW_ITEMS_UNRESOLVED or ANALYSIS_NOT_REVIEWABLE",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "snapshotVersion": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "snapshotVersion"
                ]
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "finbot_access",
        "description": "HttpOnly access cookie set by the web auth flow."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token from the native auth flow."
      }
    }
  }
}