# src/routes/onboarding.ts
import { Router } from 'express';
import type { NextFunction, Request, Response } from 'express';
import { z } from 'zod';

import { requireAuth } from '../middleware/require-auth.js';
import { validateBody } from '../middleware/validate.js';
import {
  applyReviewItemAction,
  requestRecompute,
} from '../services/corrections.service.js';
import {
  confirmFinancialReview,
  declareLinkingComplete,
  getOnboardingStatus,
  retryAnalysis,
} from '../services/onboarding-status.service.js';
import { getFinancialReviewForUser } from '../services/review.service.js';
import {
  getUserOnboarding,
  upsertUserOnboarding,
} from '../services/user-info.service.js';
import { OnboardingError } from '../types/onboarding.js';

const router = Router();

const nonNegativeNumber = z.coerce.number().finite().nonnegative();

const dateOfBirthSchema = z
  .string()
  .trim()
  .transform((value, ctx) => {
    if (/^\d{4}-\d{2}-\d{2}$/.test(value)) {
      return value;
    }

    const match = /^(0[1-9]|1[0-2])\/(0[1-9]|[12]\d|3[01])\/((?:19|20)\d{2})$/.exec(
      value,
    );

    if (!match) {
      ctx.addIssue({
        code: 'custom',
        message: 'dateOfBirth must be YYYY-MM-DD or MM/DD/YYYY',
      });
      return z.NEVER;
    }

    return `${match[3]}-${match[1]}-${match[2]}`;
  });

export const onboardingSchema = z.object({
  fullName: z.string().trim().min(1),
  dateOfBirth: dateOfBirthSchema,
  maritalStatus: z.enum([
    'single',
    'married',
    'domestic_partnership',
    'divorced',
    'widowed',
    'prefer_not_to_say',
  ]),
  dependentsCount: z.coerce.number().int().nonnegative(),
  employmentStatus: z.enum([
    'full_time',
    'part_time',
    'self_employed',
    'unemployed',
    'retired',
    'student',
  ]),
  monthlyTakeHomeIncome: nonNegativeNumber,
  monthlyHousingCosts: nonNegativeNumber,
  monthlyFoodSpend: nonNegativeNumber,
  monthlyTransportationCosts: nonNegativeNumber,
  subscriptions: z
    .array(
      z.enum([
        'netflix',
        'disney_hulu',
        'amazon_prime',
        'paramount',
        'apple',
        'xbox',
        'playstation',
        'nintendo',
        'none',
      ]),
    )
    .min(1),
  monthlyEntertainmentSubscriptionsCosts: nonNegativeNumber.optional(),
  savingsAndEmergencyFunds: nonNegativeNumber,
  totalDebt: nonNegativeNumber,
  factorInDebtInterest: z.boolean(),
  financialGoals: z
    .array(
      z.enum([
        'emergency_fund',
        'pay_off_debt',
        'save_for_retirement',
        'save_for_home',
        'invest_more',
        'reduce_spending',
      ]),
    )
    .length(3),
  additionalMoneyPools: z
    .array(
      z.enum(['vacation', 'miscellaneous', 'emergency', 'savings', 'investing']),
    )
    .length(3),
  riskComfort: z.enum(['conservative', 'moderate', 'aggressive']),
  additionalContext: z.string(),
});

/**
 * Save the manual (non-derivable) onboarding answers. Completes the manual
 * gate only — onboarding as a whole finishes when the financial review is
 * confirmed, never here. `PUT /onboarding` is kept as an alias for the
 * pre-redesign client.
 */
const saveManualHandler = [
  requireAuth,
  validateBody(onboardingSchema),
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const userId = req.user?.id;

      if (!userId) {
        res.status(401).json({ error: 'Unauthorized' });
        return;
      }

      const result = await upsertUserOnboarding(
        userId,
        req.body as z.infer<typeof onboardingSchema>,
      );

      res.status(200).json(result);
    } catch (err) {
      if (err instanceof Error && err.message.includes('violates foreign key')) {
        res.status(404).json({ error: 'User not found' });
        return;
      }

      next(err);
    }
  },
] as const;

router.put('/manual', ...saveManualHandler);
router.put('/', ...saveManualHandler);

/** Resume saved answers. 200 with `saved: null` when nothing is saved yet. */
router.get('/manual', requireAuth, async (req, res, next) => {
  try {
    const userId = req.user?.id;

    if (!userId) {
      res.status(401).json({ error: 'Unauthorized' });
      return;
    }

    const saved = await getUserOnboarding(userId);

    res.status(200).json({ saved });
  } catch (err) {
    next(err);
  }
});

/**
 * Latest financial review snapshot with coverage and actionable
 * exceptions. 409 ANALYSIS_NOT_REVIEWABLE while analysis is still running.
 */
router.get('/financial-review', requireAuth, async (req, res, next) => {
  try {
    const userId = req.user?.id;

    if (!userId) {
      res.status(401).json({ error: 'Unauthorized' });
      return;
    }

    res.status(200).json(await getFinancialReviewForUser(userId));
  } catch (err) {
    if (err instanceof OnboardingError) {
      res.status(err.statusCode).json({ error: err.message, code: err.code });
      return;
    }

    next(err);
  }
});

export const correctionSchema = z.object({
  action: z.enum([
    'accept_coverage_limitation',
    'keep_manual_value',
    'use_observed_value',
    'set_value',
    'confirm_stream',
    'dismiss_stream',
    'reclassify_transaction',
    'reclassify_merchant',
  ]),
  snapshotVersion: z.number().int().positive(),
  value: z.record(z.string(), z.unknown()).optional(),
});

/** Correct or accept one review item. */
router.patch(
  '/financial-review/items/:id',
  requireAuth,
  validateBody(correctionSchema),
  async (req, res, next) => {
    try {
      const userId = req.user?.id;

      if (!userId) {
        res.status(401).json({ error: 'Unauthorized' });
        return;
      }

      const body = req.body as z.infer<typeof correctionSchema>;

      const result = await applyReviewItemAction({
        userId,
        reviewItemId: String(req.params.id ?? ''),
        action: body.action,
        snapshotVersion: body.snapshotVersion,
        value: body.value,
      });

      res.status(200).json(result);
    } catch (err) {
      if (err instanceof OnboardingError) {
        res.status(err.statusCode).json({ error: err.message, code: err.code });
        return;
      }

      next(err);
    }
  },
);

/** Rebuild facts and review after corrections. */
router.post('/financial-review/recompute', requireAuth, async (req, res, next) => {
  try {
    const userId = req.user?.id;

    if (!userId) {
      res.status(401).json({ error: 'Unauthorized' });
      return;
    }

    const result = await requestRecompute(userId);

    res.status(202).json(result);
  } catch (err) {
    if (err instanceof OnboardingError) {
      res.status(err.statusCode).json({ error: err.message, code: err.code });
      return;
    }

    next(err);
  }
});

/** The authoritative onboarding state: gates, phase, actions, analysis. */
router.get('/status', requireAuth, async (req, res, next) => {
  try {
    const userId = req.user?.id;

    if (!userId) {
      res.status(401).json({ error: 'Unauthorized' });
      return;
    }

    res.status(200).json(await getOnboardingStatus(userId));
  } catch (err) {
    if (err instanceof OnboardingError) {
      res.status(err.statusCode).json({ error: err.message, code: err.code });
      return;
    }

    next(err);
  }
});

/**
 * The user declares they are done adding institutions. Records declared
 * coverage, backstops sync kick-off, creates the analysis run, and starts
 * analysis if every Item is already terminal.
 */
router.post('/linking-complete', requireAuth, async (req, res, next) => {
  try {
    const userId = req.user?.id;

    if (!userId) {
      res.status(401).json({ error: 'Unauthorized' });
      return;
    }

    await declareLinkingComplete(userId);

    res.status(200).json(await getOnboardingStatus(userId));
  } catch (err) {
    if (err instanceof OnboardingError) {
      res.status(err.statusCode).json({ error: err.message, code: err.code });
      return;
    }

    next(err);
  }
});

export const confirmSchema = z.object({
  snapshotVersion: z.number().int().positive(),
});

/** Confirm the latest review; the only path to onboarding completion. */
router.post(
  '/financial-review/confirm',
  requireAuth,
  validateBody(confirmSchema),
  async (req, res, next) => {
    try {
      const userId = req.user?.id;

      if (!userId) {
        res.status(401).json({ error: 'Unauthorized' });
        return;
      }

      const { snapshotVersion } = req.body as z.infer<typeof confirmSchema>;
      const result = await confirmFinancialReview(userId, snapshotVersion);

      res.status(200).json(result);
    } catch (err) {
      if (err instanceof OnboardingError) {
        res.status(err.statusCode).json({ error: err.message, code: err.code });
        return;
      }

      next(err);
    }
  },
);

/** Retry a failed analysis phase or failed institution syncs. */
router.post('/retry', requireAuth, async (req, res, next) => {
  try {
    const userId = req.user?.id;

    if (!userId) {
      res.status(401).json({ error: 'Unauthorized' });
      return;
    }

    const result = await retryAnalysis(userId);

    res
      .status(202)
      .json(
        result.status === 'already_running'
          ? { ...result, code: 'RETRY_ALREADY_QUEUED' }
          : result,
      );
  } catch (err) {
    if (err instanceof OnboardingError) {
      res.status(err.statusCode).json({ error: err.message, code: err.code });
      return;
    }

    next(err);
  }
});

export default router;